Delete `VmComputeAgent.exe`
The browser now gets upset if it detects that `VmComputeAgent.exe` is
running. After the browser enumerates the process list, it attempts to
open each image file. However, if the browser can't open the file, it acts
as if the program isn't even running. We can't just close
`VmComputeAgent.exe since it is required to run the Sandbox; however, we
can just delete the original file since the program is already running.

(Hopefully) fixes #4